pyproject.toml
[tool.poetry]
name = "tron_service"
version = "0.1.0"
description = ""
authors = ["northvenv <ulanovsega@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.12"
sqlalchemy = "^2.0.39"
tronpy = "^0.5.0"
uvicorn = "^0.34.0"
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
pytest-mock = "^3.14.0"
pydantic-settings = "^2.8.1"
alembic = "^1.15.1"
pre-commit = "^4.2.0"
ruff = "^0.11.2"
pytest-cov = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--asyncio-mode=auto"
filterwarnings = [
    "ignore::DeprecationWarning:pytest_asyncio",
    "ignore::DeprecationWarning:sqlalchemy",
    "ignore::pydantic.warnings.PydanticDeprecatedSince20"
]

[tool.ruff]
line-length = 79
target-version = "py312"

exclude = [
    "migrations/",
    "__pycache__/",
    "*.pyi",
    "tests/*",  
]

select = [
    "E",   
    "F",   
    "I001", 
    "UP",  
]

ignore = [
    "E501",  
    "F401",  
]

[tool.ruff.lint]
extend-select = [
    "B",   
    "C4",  
    "ANN", 
]

ignore = [
    "ANN101", 
    "D1",     
    "S101",  
    "ANN204",
    "B904",
    "UP035",
    "UP007",
    "ANN001",
    "ANN201",
    "UP006",
    "F811",
    "B008"
]

[tool.ruff.isort]
known-first-party = ["app"]
combine-as-imports = true

